+++ mylogging/colors/colors_module.py
@@ -9,21 +9,30 @@ from typing import Union
 # from pygments.lexers.python import PythonTracebackLexer
 # from pygments.formatters import TerminalFormatter
 
-USE_COLORS = True
 
-color_palette = {
-    "reset": "\x1b[0m",  # "reset"
-    # "INFO": "\x1b[38;21m",  # "grey"
-    # "DEBUG": "\x1b[38;21m",  # "grey"
-    "WARNING": "\x1b[33;21m",  # "yellow"
-    "ERROR": "\x1b[31;21m",  # "red"
-    "CRITICAL": "\x1b[31;1m",  # "bold_red"
-}
+class ColorsConfig:
+    USE_COLORS: bool = True
+    """You can set this variable and then use it in colorize function so you can configure it from one place.
+    It can also be configured from config module."""
+
+    COLOR_PALETTE: dict[str, str] = {
+        "reset": "\x1b[0m",  # "reset"
+        "INFO": "\x1b[38;21m",  # "grey"
+        "DEBUG": "\x1b[38;21m",  # "grey"
+        "WARNING": "\x1b[33;21m",  # "yellow"
+        "ERROR": "\x1b[31;21m",  # "red"
+        "CRITICAL": "\x1b[31;1m",  # "bold_red"
+    }
+    r"""Ansi code for defined colors like for example \x1b[33;21m"""
+
+
+colors_config = ColorsConfig()
 
 
 def colorize(message: str, level: str = "WARNING", use: Union[bool, None] = None) -> str:
-    """Add color to message based on level - usally warnings and errors, to know what is internal error on first sight.
-    There is global config.COLORIZE value that can be configured, so it's not necessary to pass as argument.
+    """Add color to message based on level - usually warnings and errors, to know what is internal error on
+    first sight. There is global config.colorize value that can be configured, so it's not necessary to pass
+    as argument.
 
     Args:
         message (str): Any string you want to color.
@@ -34,7 +43,7 @@ def colorize(message: str, level: str = "WARNING", use: Union[bool, None] = None
 
     Returns:
         str: Message in yellow color. Symbols added to string cannot be read in some terminals.
-            If config COLORIZE is False, it return original string.
+            If config colorize is False, it return original string.
 
     Example:
         >>> message = "Hello there"
@@ -42,14 +51,13 @@ def colorize(message: str, level: str = "WARNING", use: Union[bool, None] = None
         >>> colored_message
         '\\x1b[33;21m Hello there \\x1b[0m'
     """
-
     if use is None:
-        use = USE_COLORS
+        use = colors_config.USE_COLORS
 
     if not use or level in ["DEBUG", "INFO"]:
         return message
     else:
-        return f"{color_palette[level]} {message} {color_palette['reset']}"
+        return f"{colors_config.COLOR_PALETTE[level]} {message} {colors_config.COLOR_PALETTE['reset']}"
 
 
 def colorize_traceback(traceback_str: str) -> str:
@@ -68,11 +76,11 @@ def colorize_traceback(traceback_str: str) -> str:
         ...     1/0
         ... except ZeroDivisionError:
         ...     colorize_traceback(traceback.format_exc())
-        'Traceback (most recent call last):\\n  File \\x1b[36m"<doctest mylogging.colors.colorize_traceback[1]>"\\x1b[39;49;00m...'
+        'Traceback (most recent call last):\\n  File \\x1b[36m"<doctest mylogging.colors.colors...
     """
     import pygments
     from pygments.lexers.python import PythonTracebackLexer
-    from pygments.formatters import TerminalFormatter
+    from pygments.formatters import TerminalFormatter  # pylint: disable=no-name-in-module
 
     return pygments.highlight(
         traceback_str,